initial
Before, the code running on the R12 sonar head was maintained as a branch in the PyboardD repo. however, this turned out to be too confusing. Therefore, I am now moving the code to its own repo with a name that allows me to find it more quickly and maintain it without conflict with other code.